+++ extensions/archetype-base/templates/project/manifest.rb
@@ -1,4 +1,4 @@
-description "An Archetype extension that provides hooks into using HTML5 Boilerplate's CSS reset"
+description "An Archetype extension for defining the base of your page"
 
 no_configuration_file!
 skip_compilation!