+++ lib/archetype/sass_extensions/functions/styleguide/components.rb
@@ -12,23 +12,20 @@ module Archetype::SassExtensions::Styleguide
   # - {Boolean} whether or not the component was added
   #
   def styleguide_add_component(id, data, default = nil, theme = nil, force = false)
-    debug = Compass.configuration.styleguide_debug
-    helpers.debug "[archetype:styleguide] attempting to register component for `#{id}`" if debug
+    _styleguide_debug "attempting to register component for `#{id}`", :add
     _styleguide_mutex_helper(id, theme) do |id, theme|
       components = theme[:components]
       # if force was true, we have to invalidate the memoizer
       memoizer.clear(theme[:name]) if force
       # if we already have the component, don't create it again
       if component_exists?(id, theme, nil, force) || component_is_frozen?(id, theme)
-        helpers.debug "[archetype:styleguide] skipping component registration for `#{id}`. the component is already registered or frozen" if debug
+        _styleguide_debug "skipping component registration for `#{id}`. the component is already registered or frozen", :add
         return bool(false)
       end
       # otherwise add it
       components[id] = helpers.data_to_hash(default, 1, SPECIAL, ADDITIVES).merge(helpers.data_to_hash(data, 1, SPECIAL, ADDITIVES))
-      if debug
-        helpers.debug "[archetype:styleguide] successfully registered component `#{id}`"
-        helpers.debug components[id].inspect
-      end
+      _styleguide_debug "successfully registered component `#{id}`", :add
+      _styleguide_debug components[id], :add
       return bool(true)
     end
   end
@@ -49,29 +46,26 @@ module Archetype::SassExtensions::Styleguide
   # - {Boolean} whether or not the component was extended
   #
   def styleguide_extend_component(id, data, theme = nil, extension = nil, force = false)
-    debug = Compass.configuration.styleguide_debug
-    helpers.debug "[archetype:styleguide] attempting to extend component for `#{id}`" if debug
-    helpers.debug "[archetype:styleguide] extension name is `#{extension.to_sass}`" if (debug && !extension.nil?)
+    _styleguide_debug "attempting to extend component for `#{id}`", :extend
+    _styleguide_debug "extension name is `#{extension.to_sass}`", :extend unless extension.nil?
     _styleguide_mutex_helper(id, theme) do |id, theme|
       components = theme[:components]
       # if force was set, we'll create a random token for the name
       if force
         extension = random_uid('extension')
-        helpers.debug "[archetype:styleguide] forcibly extending..." if debug
+        _styleguide_debug "forcibly extending...", :extend
       end
       # use the extension name or a snapshot of the extension
       extension = helpers.to_str(extension || data.to_sass)
       extensions = theme[:extensions]
       if component_exists?(id, theme, extension, force) || component_is_frozen?(id, theme)
-        helpers.debug "[archetype:styleguide] skipping component extension for `#{id}`. the extension is already registered or frozen" if debug
+        _styleguide_debug "skipping component extension for `#{id}`. the extension is already registered or frozen", :extend
         return bool(false)
       end
       extensions.push(extension)
       components[id] = (components[id] ||= Archetype::Hash.new).rmerge(helpers.data_to_hash(data, 1, SPECIAL, ADDITIVES))
-        if debug
-          helpers.debug "[archetype:styleguide] successfully extended component `#{id}`"
-          helpers.debug components[id].inspect
-        end
+      _styleguide_debug "successfully extended component `#{id}`", :extend
+      _styleguide_debug components[id], :extend
       return bool(true)
     end
   end
@@ -111,7 +105,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_remove_component(id, theme = nil)
-    helpers.debug "[archetype:styleguide] removing component `#{id}`" if Compass.configuration.styleguide_debug
+    _styleguide_debug "removing component `#{id}`", :remove
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:components].delete(id)
       theme[:extensions].push(random_uid('remove'))
@@ -129,7 +123,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_freeze_component(id, theme = nil)
-    helpers.debug "[archetype:styleguide] freezing component `#{id}`" if Compass.configuration.styleguide_debug
+    _styleguide_debug "freezing component `#{id}`", :freeze
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:frozen] << id
       return bool(true)
@@ -145,7 +139,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_freeze_all_components(theme = nil)
-    helpers.debug "[archetype:styleguide] freezing ALL components" if Compass.configuration.styleguide_debug
+    _styleguide_debug "freezing ALL components", :freeze
     _styleguide_mutex_helper do
       theme = get_theme(theme)
       theme[:frozen] = Set.new(theme[:components].keys)
@@ -162,7 +156,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_thaw_component(id, theme = nil)
-    helpers.debug "[archetype:styleguide] thawing component `#{id}`" if Compass.configuration.styleguide_debug
+    _styleguide_debug "thawing component `#{id}`", :freeze
     _styleguide_mutex_helper(id, theme) do |id, theme|
       theme[:frozen].delete(id)
       return bool(true)
@@ -178,7 +172,7 @@ module Archetype::SassExtensions::Styleguide
   # - <tt>$theme</tt> {String} the theme to insert the component into
   #
   def styleguide_thaw_all_components(theme = nil)
-    helpers.debug "[archetype:styleguide] thawing ALL components" if Compass.configuration.styleguide_debug
+    _styleguide_debug "thawing ALL components", :freeze
     _styleguide_mutex_helper do
       theme = get_theme(theme)
       theme[:frozen].clear
@@ -199,6 +193,25 @@ module Archetype::SassExtensions::Styleguide
     theme = get_theme(theme)
     keys = theme[:components].keys.map { |k| identifier(k) }
     return list(keys, :comma)
+  end
+
+  #
+  # gets a list of all the current variants of a given component
+  #
+  # *Parameters*:
+  # - <tt>$id</tt> {String} the component identifier
+  # - <tt>$theme</tt> {String} the theme to look within
+  # *Returns*:
+  # - {List} list of component variants
+  #
+  def styleguide_component_variants(id, theme = nil)
+    id = helpers.to_str(id)
+    theme = get_theme(theme)
+    component = theme[:components][id]
+    return null if component.nil?
+    variants = component.keys
+    variants = variants.map { |k| identifier(k) }
+    return list(variants, :comma)
   end
 
   private