+++ lib/archetype/sass_extensions/functions/styleguide/grammar.rb
@@ -13,13 +13,15 @@ module Archetype::SassExtensions::Styleguide
   # - {Array} an array containing the identifer, modifiers, and a token
   #
   def grammar(sentence, theme = nil, state = nil)
-    debug = Compass.configuration.styleguide_debug
-    helpers.debug "[archetype:styleguide] converting `#{sentence}` to grammar" if debug
+    _styleguide_debug "converting `#{sentence}` to grammar", :grammar
     theme = get_theme(theme)
     components = theme[:components]
     # get a list of valid ids
     styleguideIds = components.keys
-    sentence = sentence.split if sentence.is_a? String
+
+    # convert the sentence to a string and then split into an array
+    # this ensures that all the pieces are treated as strings and not other primitive types (e.g. a list of strings in the middle of a sentence)
+    sentence = helpers.to_str(sentence).split
 
     id, modifiers = grammarize(sentence, styleguideIds)
 
@@ -32,10 +34,12 @@ module Archetype::SassExtensions::Styleguide
     # maybe in the case where we're looking for strict keys on the lookup?
     modifiers = modifiers.uniq
     token = memoizer.tokenize(theme[:name], extensions, id, modifiers, state)
-    if debug
-      helpers.debug "[archetype:styleguide] the computed grammar is..."
-      helpers.debug "  identifier: #{id}"
-      helpers.debug "  modifiers: #{modifiers}"
+    if id
+      _styleguide_debug "the computed grammar is...", :grammar
+      _styleguide_debug "  identifier: #{id}", :grammar
+      unless modifiers.empty?
+        _styleguide_debug "  modifiers: #{modifiers.join(', ')}", :grammar
+      end
     end
     return id, modifiers, token
   end